promo_service/promocode_service/requirements.txt
celery==5.2.7
Django==4.1.7
django-admin-rangefilter==0.9.0
django-extensions==3.2.1
django-import-export==3.1.0
django_debug_toolbar==3.8.1
django-stubs==1.16.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.4.1
flower==1.2.0
gunicorn==20.1.0
loguru==0.6.0
requests==2.28.2
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
python-gettext==4.1
redis==4.5.4
